requirements.txt
fastapi>=0.110,<1
uvicorn[standard]>=0.29,<1
sqlmodel>=0.0.16,<1
jinja2>=3.1,<4
python-multipart>=0.0.9,<1
bcrypt>=4.0,<6
python-jose[cryptography]>=3.3,<4
python-dotenv>=1.0,<2
aiofiles>=23.2,<25
PyQt6>=6.7,<7
PyQt6-WebEngine>=6.7,<7
